refactor move method extract class push down method
