+++ src/sqlancer/sqlite3/SQLite3Options.java
@@ -1,7 +1,5 @@
 package sqlancer.sqlite3;
 
-import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -9,18 +7,7 @@ import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 
 import sqlancer.DBMSSpecificOptions;
-import sqlancer.OracleFactory;
-import sqlancer.common.oracle.CompositeTestOracle;
-import sqlancer.common.oracle.TestOracle;
-import sqlancer.sqlite3.SQLite3Options.SQLite3OracleFactory;
-import sqlancer.sqlite3.oracle.SQLite3Fuzzer;
-import sqlancer.sqlite3.oracle.SQLite3NoRECOracle;
-import sqlancer.sqlite3.oracle.SQLite3PivotedQuerySynthesisOracle;
-import sqlancer.sqlite3.oracle.tlp.SQLite3TLPAggregateOracle;
-import sqlancer.sqlite3.oracle.tlp.SQLite3TLPDistinctOracle;
-import sqlancer.sqlite3.oracle.tlp.SQLite3TLPGroupByOracle;
-import sqlancer.sqlite3.oracle.tlp.SQLite3TLPHavingOracle;
-import sqlancer.sqlite3.oracle.tlp.SQLite3TLPWhereOracle;
+
 
 @Parameters(separators = "=", commandDescription = "SQLite3")
 public class SQLite3Options implements DBMSSpecificOptions<SQLite3OracleFactory> {
@@ -90,79 +77,79 @@ public class SQLite3Options implements DBMSSpecificOptions<SQLite3OracleFactory>
             "--execute-queries" }, description = "Specifies whether the query in the fuzzer should be executed", arity = 1)
     public boolean executeQuery = true;
 
-    public enum SQLite3OracleFactory implements OracleFactory<SQLite3GlobalState> {
-        PQS {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3PivotedQuerySynthesisOracle(globalState);
-            }
-
-            @Override
-            public boolean requiresAllTablesToContainRows() {
-                return true;
-            }
-
-        },
-        NoREC {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3NoRECOracle(globalState);
-            }
-        },
-        AGGREGATE {
-
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3TLPAggregateOracle(globalState);
-            }
-
-        },
-        WHERE {
-
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3TLPWhereOracle(globalState);
-            }
-
-        },
-        DISTINCT {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3TLPDistinctOracle(globalState);
-            }
-        },
-        GROUP_BY {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3TLPGroupByOracle(globalState);
-            }
-        },
-        HAVING {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3TLPHavingOracle(globalState);
-            }
-        },
-        FUZZER {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                return new SQLite3Fuzzer(globalState);
-            }
-        },
-        QUERY_PARTITIONING {
-            @Override
-            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
-                List<TestOracle> oracles = new ArrayList<>();
-                oracles.add(new SQLite3TLPWhereOracle(globalState));
-                oracles.add(new SQLite3TLPDistinctOracle(globalState));
-                oracles.add(new SQLite3TLPGroupByOracle(globalState));
-                oracles.add(new SQLite3TLPHavingOracle(globalState));
-                oracles.add(new SQLite3TLPAggregateOracle(globalState));
-                return new CompositeTestOracle(oracles, globalState);
-            }
-        };
-
-    }
+//    public enum SQLite3OracleFactory implements OracleFactory<SQLite3GlobalState> {
+//        PQS {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3PivotedQuerySynthesisOracle(globalState);
+//            }
+//
+//            @Override
+//            public boolean requiresAllTablesToContainRows() {
+//                return true;
+//            }
+//
+//        },
+//        NoREC {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3NoRECOracle(globalState);
+//            }
+//        },
+//        AGGREGATE {
+//
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3TLPAggregateOracle(globalState);
+//            }
+//
+//        },
+//        WHERE {
+//
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3TLPWhereOracle(globalState);
+//            }
+//
+//        },
+//        DISTINCT {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3TLPDistinctOracle(globalState);
+//            }
+//        },
+//        GROUP_BY {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3TLPGroupByOracle(globalState);
+//            }
+//        },
+//        HAVING {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3TLPHavingOracle(globalState);
+//            }
+//        },
+//        FUZZER {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                return new SQLite3Fuzzer(globalState);
+//            }
+//        },
+//        QUERY_PARTITIONING {
+//            @Override
+//            public TestOracle create(SQLite3GlobalState globalState) throws SQLException {
+//                List<TestOracle> oracles = new ArrayList<>();
+//                oracles.add(new SQLite3TLPWhereOracle(globalState));
+//                oracles.add(new SQLite3TLPDistinctOracle(globalState));
+//                oracles.add(new SQLite3TLPGroupByOracle(globalState));
+//                oracles.add(new SQLite3TLPHavingOracle(globalState));
+//                oracles.add(new SQLite3TLPAggregateOracle(globalState));
+//                return new CompositeTestOracle(oracles, globalState);
+//            }
+//        };
+//
+//    }
 
     @Override
     public List<SQLite3OracleFactory> getTestOracleFactory() {
